requirements.txt
# 智能巡检系统 - API服务端依赖

# BentoML服务框架
bentoml==1.4.28

# 数据处理
numpy==2.3.4
Pillow==12.0.0

# 大模型客户端
openai==2.7.1
python-dotenv==1.2.1

# 配置管理
PyYAML==6.0.3

# Web框架（未来Web服务使用）
Flask==3.0.0
Flask-SocketIO==5.3.5
Flask-CORS==4.0.0

# 数据验证
pydantic==2.12.4
pydantic-core==2.41.5

# 工具库
requests==2.32.5

